backend/requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.0.5
django-cors-headers==4.3.1
djangorestframework==3.15.1
sqlparse==0.5.0
